requirements.txt
accelerate==0.25.0
decord==0.6.0
einops==0.7.0
fire==0.5.0
imageio==2.34.0
# ipython==8.12.3
ipython==8.18.1
jaxtyping==0.2.28
kmeans_gpu==0.0.5
matplotlib==3.7.2
mediapy==1.2.0
numpy==1.24.2
omegaconf==2.1.1
open3d==0.18.0
opencv_python==4.6.0.66
opencv_python_headless==4.9.0.80
# Pillow==9.5.0
Pillow==10.3.0
plyfile==1.0.3
point_cloud_utils==0.30.2
pyfqmr==0.2.0
pygltflib==1.16.2
PyMCubes==0.1.4
pymeshlab==2023.12
safetensors==0.3.3
scikit_learn==1.3.2
scipy==1.13.0
# simple_knn==0.0.0
# pip install -e git+https://gitlab.inria.fr/bkerbl/simple-knn.git@44f764299fa305faf6ec5ebd99939e0508331503#egg=simple_knn&subdirectory=../../../submodules/simple-knn
# torch==2.2.2+cu121 # pip install torch==2.2.2 torchvision==0.17.2 torchaudio==2.2.2 --index-url https://download.pytorch.org/whl/cu121
# torchvision==0.17.2
tqdm==4.65.0
trimesh==4.0.8
warp_lang==0.10.1
xatlas==0.0.9
wandb
# pip install git+https://github.com/graphdeco-inria/diff-gaussian-rasterization


# for sgm compatibility
# pytorch-lightning==2.4.0
# kornia==0.8.0 
# open-clip-torch==2.30.0
# transformers==4.19.1